feat: add query param to cache bust image urls
Currently, when a background is approved, if a user has recently had
another image approved Discord will show the old background in the
"Approved" embded due to its media cache. Additionally, `.gif` is added
to the end of the URL so Discord will treat the image as a gif when
displaying in an embed (because Discord is incredibly big brain and
ignores mime types and uses file extensions instead for gif detection).